services-data-fetching: Move request to page, add error handling, conditional render
