+++ src/components/common/services-list/services-list.tsx
@@ -1,10 +1,7 @@
-import { getServices } from '@/../sanity/api';
 import { ServiceCard } from '../service-card/service-card';
-import { ServiceData } from './types';
+import { ServiceData, ServicesProps } from '../../../sections/services/types';
 
-export const ServicesList = async () => {
-  const services = await getServices();
-
+export const ServicesList = async ({ services }: ServicesProps) => {
   return (
     <ul className="flex flex-col md:flex-row flex-wrap gap-8 md:gap-x-[94px] md:gap-y-[60px] xl:gap-x-[125px]">
       {services.map((service: ServiceData) => (
