+++ src/sections/services/types.ts
@@ -2,3 +2,7 @@ export type ServiceData = {
   _key: string;
   title: string;
 };
+
+export type ServicesProps = {
+  services: ServiceData[];
+};
